django-ml-commune/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
pandas==1.5.3
scikit-learn==1.2.2
numpy==1.24.2
matplotlib==3.7.1
joblib==1.2.0
python-dotenv==1.0.0